# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "1234567890123456",
                      "email": "user@example.com",
                      "firstName": "John",
                      "lastName": "Doe",
                      "organisation": "Acme Corp",
                      "role": "DEVELOPER",
                      "isVerified": false,
                      "createdAt": "2024-01-01T00:00:00.000Z"
                    },
                    "apiKey": "esp_1234567890abcdef...",
                    "wallets": {
                      "custodial": {
                        "created": true,
                        "count": 3,
                        "networks": [
                          "Circles"
                        ]
                      },
                      "circle": {
                        "status": "success",
                        "created": true,
                        "count": 2,
                        "blockchains": [
                          "ETH-SEPOLIA",
                          "MATIC-AMOY"
                        ],
                        "configured": true
                      }
                    },
                    "message": "User created successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Register new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1...",
                    "user": {
                      "id": "1234567890123456",
                      "email": "user@example.com",
                      "firstName": "John",
                      "lastName": "Doe",
                      "organisation": "Acme Corp",
                      "role": "DEVELOPER",
                      "walletAddress": "0x..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Login user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid OTP"
          }
        },
        "summary": "Verify email with OTP",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/request-password-reset": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset instructions sent"
          }
        },
        "summary": "Request password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid OTP or password"
          }
        },
        "summary": "Reset password with OTP",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1234567890123456",
                    "email": "user@example.com",
                    "firstName": "John",
                    "lastName": "Doe",
                    "organisation": "Acme Corp",
                    "role": "DEVELOPER"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user profile",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Change password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/profile": {
      "get": {
        "operationId": "ProfileController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "firstName": "John",
                    "lastName": "Doe",
                    "email": "john@example.com",
                    "phone": "+1 234 567 8900",
                    "organisation": "Acme Corp",
                    "kycLevel": "Level 1",
                    "transactionLimit": "$1,000/day",
                    "apiKey": "********",
                    "twoFactorEnabled": false,
                    "apiAccess": true,
                    "webhookNotifications": false
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user profile details",
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "EspeePay API",
    "description": "The EspeePay API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "Test123!@#",
            "description": "User password (min 8 characters)"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "country": {
            "type": "string",
            "example": "US",
            "description": "User country"
          },
          "organisation": {
            "type": "string",
            "example": "Acme Corp",
            "description": "User organization"
          },
          "role": {
            "type": "string",
            "enum": [
              "BUSINESS",
              "DEVELOPER"
            ],
            "example": "DEVELOPER",
            "description": "User role (BUSINESS or DEVELOPER)"
          },
          "referralCode": {
            "type": "string",
            "description": "Optional referral code"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "country",
          "organisation",
          "role"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "Test123!@#",
            "description": "User password (min 8 characters)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "6-digit OTP code"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "6-digit OTP code"
          },
          "newPassword": {
            "type": "string",
            "example": "NewTest123!@#",
            "description": "New password (min 8 characters)"
          }
        },
        "required": [
          "email",
          "otp",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  }
}